+++ packages/central-server/__tests__/utilities.js
@@ -8,7 +8,7 @@ import { fake } from '@tamanu/fake-data/fake';
 import { asNewRole } from '@tamanu/shared/test-helpers';
 import { sleepAsync } from '@tamanu/utils/sleepAsync';
 import { initReporting } from '@tamanu/database/services/reporting';
-import { initFhirSettingsFromDb } from '@tamanu/shared/utils/fhir/fhirSettingsCache';
+import { initFhirSettingsFromDb } from '@tamanu/shared/utils/fhir/fhirSettings';
 
 import { buildToken } from '../dist/auth/utils';
 import { createApp } from '../dist/createApp';
