+++ packages/central-server/app/database/index.js
@@ -2,7 +2,7 @@ import config from 'config';
 
 import { addHooks } from './hooks';
 import { closeAllDatabases, openDatabase } from '@tamanu/database/services/database';
-import { initFhirSettingsFromDb } from '@tamanu/shared/utils/fhir/fhirSettingsCache';
+import { initFhirSettingsFromDb } from '@tamanu/shared/utils/fhir/fhirSettings';
 import { setFhirRefreshTriggers } from './setFhirRefreshTriggers';
 
 const getOrCreateConnection = async ({ testMode, ...configOverrides }, key = 'main') => {
