+++ packages/shared/src/utils/fhir/index.js
@@ -1,6 +1,6 @@
 export * from './datetime';
 export * from './errors';
-export * from './fhirSettingsCache';
+export * from './fhirSettings';
 export * from './parameters';
 export * from './schemata';
 export * from './OperationOutcome';
